And now we have a proper ScreenSaver!
Converted from standard Windows .exe (WinMain) to a proper Screen Server
(ScreenSaverProc). This gives us:

- No more random exits on WinXP.
- The mini-preview works!

The included scrnsave.c file is taken from:

http://sourceforge.net/p/mingw-w64/mingw-w64/ci/master/tree/mingw-w64-crt/libsrc/scrnsave.c

I had to add a local copy, because the because the original one is wrapped in
one big "#if 0" for some reason, and thus MinGW64's libscrnsave.a, contains no
no code at all.

Only change to that file was commenting out said "#if 0".